Clean up front end
