+++ app/web/.cursorrules
@@ -40,4 +40,5 @@ To run the app locally in development, use `yarn start` (not Docker)
 - Use default Material-UI hover styles instead of custom dark grey on hover
 - Prefer modern 2025 design patterns for UI components
 - Use StyledLink.tsx component or import from next/link for links to preserve routing, NOT MUI Link
+- Always put types at the top of the file below the imports
