+++ app/web/features/auth/locales/en.json
@@ -322,13 +322,14 @@
   },
   "volunteer_management": {
     "title": "Volunteer Management",
-    "description": "Manage your volunteer profile information. This controls how you appear on the team page. If your role or volunteering dates are incorrect, please let us know and we'll update them.",
+    "description": "If your role or dates are incorrect, please let us know.",
+    "card_preview_label": "Your team page listing",
     "not_a_volunteer_message": "According to our records you are not a current or past volunteer. If this is incorrect, please let us know by filling in <1>this form</1>.",
     "current_volunteer": "Volunteering since {{startDate}}.",
     "past_volunteer": "Volunteered from {{startDate}} to {{endDate}}.",
     "shown_on_team_page_current": "Shown on team page as a current team member.",
     "shown_on_team_page_past": "Shown on team page as a past volunteer.",
-    "not_shown_on_team_page": "Not shown on team page.",
+    "not_shown_on_team_page": "Hidden from team page.",
     "update_success": "Your volunteer information has been updated.",
     "form": {
       "title": "Update your volunteer info",
@@ -343,15 +344,15 @@
       "link_type_linkedin": "LinkedIn",
       "link_type_website": "Website",
       "link_text_label": "Link value",
-      "link_text_linkedin_placeholder": "your-linkedin-username",
+      "link_text_linkedin_placeholder": "jane-doe-a1b2c3",
       "link_text_email_placeholder": "your@email.com",
       "link_text_website_placeholder": "https://your-website.com",
-      "link_text_linkedin_helper": "Your LinkedIn username (without the URL)",
+      "link_text_linkedin_helper": "The username from your profile URL (e.g. linkedin.com/in/jane-doe-a1b2c3)",
       "link_text_email_helper": "Your email address",
       "link_text_website_helper": "The text to display for the link",
       "link_text_required": "Please enter a value for the link",
       "link_text_email_invalid": "Please enter a valid email address",
-      "link_text_linkedin_invalid": "Please enter a valid LinkedIn profile URL",
+      "link_text_linkedin_invalid": "Please enter a valid LinkedIn username (letters, numbers, and hyphens only)",
       "link_url_label": "Link URL",
       "link_url_placeholder": "https://your-website.com",
       "link_url_helper": "The full URL the link should point to",
