+++ app/web/features/auth/volunteer/useVolunteerInfo.ts
@@ -1,0 +1,13 @@
+import { useQuery } from "@tanstack/react-query";
+import { volunteerInfoQueryKey } from "features/queryKeys";
+import { RpcError } from "grpc-web";
+import { GetMyVolunteerInfoRes } from "proto/account_pb";
+import { service } from "service";
+
+export function useVolunteerInfo(enabled = true) {
+  return useQuery<GetMyVolunteerInfoRes.AsObject, RpcError>({
+    queryKey: [volunteerInfoQueryKey],
+    queryFn: service.account.getMyVolunteerInfo,
+    enabled,
+  });
+}
